Fusion centroid

centroid_fusion calculates the centroid of each timestep in fusion events. The function expects an edge-list of fusion events identified by fusion_id() from edge-lists generated with edge_dist() and a data.table with relocation data appended with a timegroup column from group_times(). Relocation data should be in two columns representing the X and Y coordinates, or in a geometry column prepared by the helper function get_geometry().

Usage

centroid_fusion(
  edges = NULL,
  DT = NULL,
  id = NULL,
  coords = NULL,
  crs = NULL,
  timegroup = "timegroup",
  geometry = "geometry"
)

Arguments

edges

edge-list generated generated by edge_dist() or edge_nn(), with fusionID column generated by fusion_id()

DT

input data.table with timegroup column generated with group_times() matching the input data.table used to generate the edge list with edge_nn() or edge_dist()

id

character string of ID column name

coords

character vector of X coordinate and Y coordinate column names. Note: the order is assumed X followed by Y column names

crs

numeric or character defining the coordinate reference system to be passed to sf::st_crs. For example, either crs = "EPSG:32736" or crs = 32736. Used only if coords are provided, see details under Interface

timegroup

timegroup field in the DT within which the output will be calculated

geometry

simple feature geometry list column name, generated by get_geometry(). Default 'geometry', see details under Interface

Value

centroid_fusion returns the input edges appended with centroid column(s) for each timestep and fusion id.

If coords are provided, the centroid columns will be named by prefixing the coordinate column names with "centroid_" (eg. "X" = "centroid_X"). If geometry is used, the centroid column will be named "centroid".

Note: due to the merge required within this function, the output needs to be reassigned unlike some other spatsoc functions like dyad_id and group_pts. See details in FAQ.

A message is returned when the centroid column(s) already exist in the input because they will be overwritten.

Details

The edges and DT must be data.tables. If your data is a data.frame, you can convert it by reference using data.table::setDT() or by reassigning using data.table::data.table().

The edges and DT are internally merged in this function using the columns timegroup (from group_times()) and ID1 and ID2 (in edges, from dyad_id) and id (in DT). This function expects a fusionID present, generated with the fusion_id() function. The timegroup argument expects the names of a column in edges which correspond to the timegroup column. The id and timegroup arguments expect the names of columns in DT which correspond to the id, and timegroup columns.

See below under "Interface" for details on providing coordinates and under "Centroid function" for details on the underlying centroid function used.

Interface

Two interfaces are available for providing coordinates:

Provide coords and optionally crs. The coords argument expects the names of the X and Y coordinate columns. The crs argument expects a character string or numeric defining the coordinate reference system to be passed to sf::st_crs. For example, for UTM zone 36S (EPSG 32736), the crs argument is crs = "EPSG:32736" or crs = 32736. See https://spatialreference.org for a list of EPSG codes. For centroid calculations, if crs is NULL, it will be internally set to NA_crs_.

(New!) Provide geometry. The geometry argument allows the user to supply a geometry column that represents the coordinates as a simple feature geometry list column. This interface expects the user to prepare their input DT with get_geometry(). To use this interface, leave the coords and crs arguments NULL, and the default argument for geometry ('geometry') will be used directly.

Centroid function

The underlying centroid function used depends on the crs of the coordinates or geometry provided.

If the crs is longlat degrees (as determined by sf::st_is_longlat()) and sf::sf_use_s2() is TRUE, the distance function is sf::st_centroid() which passes to s2::s2_centroid().

If the crs is longlat degrees but sf::sf_use_s2() is FALSE, the centroid calculated will be incorrect. See sf::st_centroid().

If the crs is not longlat degrees (eg. NULL, NA_crs_, or projected), the centroid function used is mean.

Note: if the input is length 1, the input is returned.

See also

Examples

# Load data.table
library(data.table)
# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))
# Cast the character column to POSIXct
DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]
#>          ID        X       Y            datetime population
#>      <char>    <num>   <num>              <POSc>      <int>
#>   1:      A 696191.5 5508362 2017-01-17 00:00:47          1
#>   2:      A 696205.2 5508363 2017-01-17 02:00:48          1
#>   3:      A 696745.8 5508225 2017-01-17 04:00:48          1
#>   4:      A 696952.0 5508373 2017-01-17 06:00:54          1
#>   5:      A 696079.0 5508218 2017-01-17 08:00:54          1
#>  ---                                                       
#> 116:      J 696996.5 5508024 2017-01-17 14:00:42          1
#> 117:      J 697046.4 5507922 2017-01-17 16:00:47          1
#> 118:      J 697037.5 5507924 2017-01-17 18:00:54          1
#> 119:      J 697303.0 5508347 2017-01-17 20:00:24          1
#> 120:      J 696616.7 5508736 2017-01-17 22:00:42          1

# Temporal grouping
group_times(DT, datetime = 'datetime', threshold = '20 minutes')
#>          ID        X       Y            datetime population minutes timegroup
#>      <char>    <num>   <num>              <POSc>      <int>   <int>     <int>
#>   1:      A 696191.5 5508362 2017-01-17 00:00:47          1       0         1
#>   2:      A 696205.2 5508363 2017-01-17 02:00:48          1       0         2
#>   3:      A 696745.8 5508225 2017-01-17 04:00:48          1       0         3
#>   4:      A 696952.0 5508373 2017-01-17 06:00:54          1       0         4
#>   5:      A 696079.0 5508218 2017-01-17 08:00:54          1       0         5
#>  ---                                                                         
#> 116:      J 696996.5 5508024 2017-01-17 14:00:42          1       0         8
#> 117:      J 697046.4 5507922 2017-01-17 16:00:47          1       0         9
#> 118:      J 697037.5 5507924 2017-01-17 18:00:54          1       0        10
#> 119:      J 697303.0 5508347 2017-01-17 20:00:24          1       0        11
#> 120:      J 696616.7 5508736 2017-01-17 22:00:42          1       0        12

# Edge-list generation
edges <- edge_dist(
    DT,
    threshold = 100,
    id = 'ID',
    coords = c('X', 'Y'),
    timegroup = 'timegroup',
    returnDist = TRUE,
    fillNA = FALSE
  )

# Generate dyad id
dyad_id(edges, id1 = 'ID1', id2 = 'ID2')
#>      timegroup    ID1    ID2  distance dyadID
#>          <int> <char> <char>     <num> <char>
#>   1:         1      C      A 73.474248    A-C
#>   2:         1      E      A 29.379813    A-E
#>   3:         1      F      A 26.212467    A-F
#>   4:         1      I      A 22.694875    A-I
#>   5:         1      G      B 20.476215    B-G
#>  ---                                         
#> 312:        12      J      I 14.453578    I-J
#> 313:        12      A      J 98.478792    A-J
#> 314:        12      C      J 84.786530    C-J
#> 315:        12      E      J  4.272393    E-J
#> 316:        12      I      J 14.453578    I-J

# Generate fusion id
fusion_id(edges, threshold = 100)
#>      timegroup    ID1    ID2  distance dyadID fusionID
#>          <int> <char> <char>     <num> <char>    <int>
#>   1:         1      C      A 73.474248    A-C        1
#>   2:         1      E      A 29.379813    A-E        2
#>   3:         1      F      A 26.212467    A-F        3
#>   4:         1      I      A 22.694875    A-I        4
#>   5:         1      G      B 20.476215    B-G        5
#>  ---                                                  
#> 312:        12      J      I 14.453578    I-J       24
#> 313:        12      A      J 98.478792    A-J       18
#> 314:        12      C      J 84.786530    C-J       20
#> 315:        12      E      J  4.272393    E-J       14
#> 316:        12      I      J 14.453578    I-J       24

# Calculate fusion centroid
centroids <- centroid_fusion(
  edges,
  DT,
  id = 'ID',
  coords = c('X', 'Y'),
  timegroup = 'timegroup'
)

print(centroids)
#>      timegroup    ID1    ID2  distance dyadID fusionID centroid_X centroid_Y
#>          <int> <char> <char>     <num> <char>    <int>      <num>      <num>
#>   1:         1      C      A 73.474248    A-C        1   696218.5    5508337
#>   2:         1      E      A 29.379813    A-E        2   696201.1    5508351
#>   3:         1      F      A 26.212467    A-F        3   696203.6    5508357
#>   4:         1      I      A 22.694875    A-I        4   696199.9    5508354
#>   5:         1      G      B 20.476215    B-G        5   697691.9    5511758
#>  ---                                                                        
#> 312:        12      J      I 14.453578    I-J       24   696621.7    5508731
#> 313:        12      A      J 98.478792    A-J       18   696640.7    5508693
#> 314:        12      C      J 84.786530    C-J       20   696617.8    5508693
#> 315:        12      E      J  4.272393    E-J       14   696617.8    5508734
#> 316:        12      I      J 14.453578    I-J       24   696621.7    5508731

# Or, using the new geometry interface
get_geometry(DT, coords = c('X', 'Y'), crs = 32736)
#>          ID        X       Y            datetime population minutes timegroup
#>      <char>    <num>   <num>              <POSc>      <int>   <int>     <int>
#>   1:      A 696191.5 5508362 2017-01-17 00:00:47          1       0         1
#>   2:      A 696205.2 5508363 2017-01-17 02:00:48          1       0         2
#>   3:      A 696745.8 5508225 2017-01-17 04:00:48          1       0         3
#>   4:      A 696952.0 5508373 2017-01-17 06:00:54          1       0         4
#>   5:      A 696079.0 5508218 2017-01-17 08:00:54          1       0         5
#>  ---                                                                         
#> 116:      J 696996.5 5508024 2017-01-17 14:00:42          1       0         8
#> 117:      J 697046.4 5507922 2017-01-17 16:00:47          1       0         9
#> 118:      J 697037.5 5507924 2017-01-17 18:00:54          1       0        10
#> 119:      J 697303.0 5508347 2017-01-17 20:00:24          1       0        11
#> 120:      J 696616.7 5508736 2017-01-17 22:00:42          1       0        12
#>                      geometry
#>                   <sfc_POINT>
#>   1: POINT (696191.5 5508362)
#>   2: POINT (696205.2 5508363)
#>   3: POINT (696745.8 5508225)
#>   4:   POINT (696952 5508373)
#>   5:   POINT (696079 5508218)
#>  ---                         
#> 116: POINT (696996.5 5508024)
#> 117: POINT (697046.4 5507922)
#> 118: POINT (697037.5 5507924)
#> 119:   POINT (697303 5508347)
#> 120: POINT (696616.7 5508736)
edges <- edge_dist(DT, threshold = 100, id = 'ID', timegroup = 'timegroup', returnDist = TRUE)
dyad_id(edges, id = 'ID1', id2 = 'ID2')
#> Key: <timegroup, ID1>
#>      timegroup    ID1    ID2  distance dyadID
#>          <int> <char> <char>     <num> <char>
#>   1:         1      A      C 73.474248    A-C
#>   2:         1      A      E 29.379813    A-E
#>   3:         1      A      F 26.212467    A-F
#>   4:         1      A      I 22.694875    A-I
#>   5:         1      B      G 20.476215    B-G
#>  ---                                         
#> 339:        12      I      J 14.453578    I-J
#> 340:        12      J      A 98.478792    A-J
#> 341:        12      J      C 84.786530    C-J
#> 342:        12      J      E  4.272393    E-J
#> 343:        12      J      I 14.453578    I-J
fusion_id(edges, threshold = 100)
#> Key: <timegroup, ID1>
#>      timegroup    ID1    ID2  distance dyadID fusionID
#>          <int> <char> <char>     <num> <char>    <int>
#>   1:         1      A      C 73.474248    A-C        1
#>   2:         1      A      E 29.379813    A-E        2
#>   3:         1      A      F 26.212467    A-F        3
#>   4:         1      A      I 22.694875    A-I        4
#>   5:         1      B      G 20.476215    B-G        5
#>  ---                                                  
#> 339:        12      I      J 14.453578    I-J       24
#> 340:        12      J      A 98.478792    A-J       18
#> 341:        12      J      C 84.786530    C-J       20
#> 342:        12      J      E  4.272393    E-J       14
#> 343:        12      J      I 14.453578    I-J       24
centroids <- centroid_fusion(
  edges,
  DT,
  id = 'ID',
  timegroup = 'timegroup'
)
print(centroids)
#>      timegroup    ID1    ID2  distance dyadID fusionID                 geometry
#>          <int> <char> <char>     <num> <char>    <int>              <sfc_POINT>
#>   1:         1      A      C 73.474248    A-C        1 POINT (696191.5 5508362)
#>   2:         1      A      E 29.379813    A-E        2 POINT (696191.5 5508362)
#>   3:         1      A      F 26.212467    A-F        3 POINT (696191.5 5508362)
#>   4:         1      A      I 22.694875    A-I        4 POINT (696191.5 5508362)
#>   5:         1      B      G 20.476215    B-G        5 POINT (697682.7 5511754)
#>  ---                                                                           
#> 339:        12      I      J 14.453578    I-J       24 POINT (696626.8 5508725)
#> 340:        12      J      A 98.478792    A-J       18 POINT (696616.7 5508736)
#> 341:        12      J      C 84.786530    C-J       20 POINT (696616.7 5508736)
#> 342:        12      J      E  4.272393    E-J       14 POINT (696616.7 5508736)
#> 343:        12      J      I 14.453578    I-J       24 POINT (696616.7 5508736)
#>                      centroid
#>                   <sfc_POINT>
#>   1: POINT (696218.5 5508337)
#>   2: POINT (696201.1 5508351)
#>   3: POINT (696203.6 5508357)
#>   4: POINT (696199.9 5508354)
#>   5: POINT (697691.9 5511758)
#>  ---                         
#> 339: POINT (696621.7 5508731)
#> 340: POINT (696640.7 5508693)
#> 341: POINT (696617.8 5508693)
#> 342: POINT (696617.8 5508734)
#> 343: POINT (696621.7 5508731)